+++ crates/voicevox_core_python_api/python/voicevox_core/__init__.py
@@ -25,29 +25,26 @@ from ._rust import (  # noqa: F401
     ModelAlreadyLoadedError,
     ModelNotFoundError,
     NotLoadedOpenjtalkDictError,
-    OpenJtalk,
     OpenZipFileError,
     ParseKanaError,
     ReadZipEntryError,
     SaveUserDictError,
     StyleAlreadyLoadedError,
     StyleNotFoundError,
-    Synthesizer,
-    UserDict,
     UseUserDictError,
-    VoiceModel,
     WordNotFoundError,
     __version__,
     supported_devices,
 )
 
-from . import blocking  # noqa: F401 isort: skip
+from . import asyncio, blocking  # noqa: F401 isort: skip
 
 __all__ = [
     "__version__",
     "AccelerationMode",
     "AccentPhrase",
     "AudioQuery",
+    "asyncio",
     "blocking",
     "ExtractFullContextLabelError",
     "GetSupportedDevicesError",
@@ -60,7 +57,6 @@ __all__ = [
     "ModelNotFoundError",
     "Mora",
     "NotLoadedOpenjtalkDictError",
-    "OpenJtalk",
     "OpenZipFileError",
     "ParseKanaError",
     "ReadZipEntryError",
@@ -71,11 +67,8 @@ __all__ = [
     "StyleNotFoundError",
     "StyleVersion",
     "SupportedDevices",
-    "Synthesizer",
-    "VoiceModel",
     "supported_devices",
     "UseUserDictError",
-    "UserDict",
     "UserDictWord",
     "UserDictWordType",
     "VoiceModelId",
