+++ crates/voicevox_core_python_api/python/voicevox_core/asyncio.py
@@ -1,0 +1,3 @@
+from ._rust.asyncio import OpenJtalk, Synthesizer, UserDict, VoiceModel
+
+__all__ = ["OpenJtalk", "Synthesizer", "UserDict", "VoiceModel"]
